# server/docs/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: SafeVideo KYC API
  description: |
    SafeVideo KYC (Know Your Customer) システムのAPIドキュメント
    
    ## 認証
    このAPIは2種類の認証方式をサポートしています：
    - **JWT認証**: `Authorization: Bearer <token>`
    - **Firebase認証**: `Firebase-Token: <token>`
    
    ## レート制限
    APIエンドポイントには以下のレート制限が適用されます：
    - デフォルト: 15分間に100リクエスト
    - バッチAPI: 15分間に10リクエスト
    - 検索API: 1分間に30リクエスト
    
    ## エラーレスポンス
    すべてのエラーレスポンスは統一された形式で返されます。
  version: 1.0.0
  contact:
    name: SafeVideo Team
    email: REPLACE_WITH_ACTUAL_ADDRESS
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 開発サーバー
  - url: https://api.safevideo.com/v1
    description: 本番サーバー
paths:
...
  /batch/performers:
    post:
      tags:
        - Batch
      summary: パフォーマー一括インポート
      description: |
        パフォーマーデータを一括でインポートします。
        JSONデータまたはCSVファイルでのインポートに対応しています。
      security:
        - bearerAuth: []
        - firebaseAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                performers:
                  type: array
                  items:
                    $ref: '#/components/schemas/PerformerInput'
                dryRun:
                  type: boolean
                  default: false
                  description: テストモード（実際のデータ更新は行わない）
                skipDuplicates:
                  type: boolean
                  default: true
                  description: 重複データをスキップ
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSVファイル
                dryRun:
                  type: boolean
                  default: false
      responses:
        '202':
          description: バッチジョブが開始されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
        '400':
          description: リクエストが無効です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 認証が必要です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: アクセス権限がありません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /batch/jobs/{jobId}:
    get:
      tags:
        - Batch
      summary: バッチジョブ状況取得
      description: 指定されたバッチジョブの実行状況を取得します
      security:
        - bearerAuth: []
        - firebaseAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
          description: バッチジョブID
      responses:
        '200':
          description: ジョブ情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJob'
        '404':
          description: ジョブが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Batch
      summary: バッチジョブキャンセル
      description: 実行中のバッチジョブをキャンセルします
      security:
        - bearerAuth: []
        - firebaseAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
          description: バッチジョブID
      responses:
        '200':
          description: ジョブがキャンセルされました
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "ジョブをキャンセルしました"
        '400':
          description: キャンセルできないジョブです
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Bulk API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    firebaseAuth:
      type: apiKey
      in: header
      name: Firebase-Token
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: エラーメッセージ
            statusCode:
              type: integer
              description: HTTPステータスコード
            details:
              type: object
              description: エラー詳細情報
            timestamp:
              type: string
              format: date-time
              description: エラー発生時刻
      example:
        error:
          code: "VALIDATION_ERROR"
          message: "入力データが無効です"
          statusCode: 400
          details:
            errors:
              - field: "email"
                message: "有効なメールアドレスを入力してください"
          timestamp: "2024-01-01T12:00:00.000Z"
    PerformerInput:
      type: object
      properties:
        lastName:
          type: string
          description: 姓
        firstName:
          type: string
          description: 名
        lastNameRoman:
          type: string
          description: 姓（ローマ字）
        firstNameRoman:
          type: string
          description: 名（ローマ字）
        email:
          type: string
          format: email
          description: メールアドレス
        phone:
          type: string
          description: 電話番号
        birthDate:
          type: string
          format: date
          description: 生年月日
        nationality:
          type: string
          description: 国籍
        address:
          type: string
          description: 住所
        notes:
          type: string
          description: 備考
      required:
        - lastName
        - firstName
    BatchJobResponse:
      type: object
      properties:
        jobId:
          type: string
          description: バッチジョブID
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
          description: ジョブステータス
        totalItems:
          type: integer
          description: 処理対象アイテム数
        message:
          type: string
          description: メッセージ
        estimatedTime:
          type: string
          description: 推定完了時間
      example:
        jobId: "batch_123456789"
        status: "processing"
        totalItems: 100
        message: "バッチ処理を開始しました"
        estimatedTime: "2-3 minutes"
    BatchJob:
      type: object
      properties:
        id:
          type: string
          description: ジョブID
        userId:
          type: integer
          description: 実行ユーザーID
        jobType:
          type: string
          description: ジョブタイプ
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        totalItems:
          type: integer
          description: 総アイテム数
        processedItems:
          type: integer
          description: 処理済みアイテム数
        successItems:
          type: integer
          description: 成功アイテム数
        failedItems:
          type: integer
          description: 失敗アイテム数
        progress:
          type: number
          minimum: 0
          maximum: 100
          description: 進捗率（%）
        errorMessage:
          type: string
          description: エラーメッセージ
        result:
          type: object
          description: 処理結果
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...